refactor: split lint command responsibilities
