+++ cmd/rslint/lint_pipeline.go
@@ -3,380 +3,31 @@ package main
 import (
 	"cmp"
 	"context"
-	"errors"
-	"flag"
 	"fmt"
 	"os"
-	"path/filepath"
 	"runtime"
 	"runtime/pprof"
 	"runtime/trace"
 	"slices"
-	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
-
-	"github.com/web-infra-dev/rslint/internal/linter"
-	"github.com/web-infra-dev/rslint/internal/output"
-	"github.com/web-infra-dev/rslint/internal/program/loader"
-	"github.com/web-infra-dev/rslint/internal/rule"
-	"github.com/web-infra-dev/rslint/internal/rules"
-	"github.com/web-infra-dev/rslint/internal/utils"
 
 	"github.com/microsoft/typescript-go/shim/ast"
 	"github.com/microsoft/typescript-go/shim/bundled"
 	"github.com/microsoft/typescript-go/shim/core"
 	"github.com/microsoft/typescript-go/shim/tspath"
-	"github.com/microsoft/typescript-go/shim/vfs"
 	"github.com/microsoft/typescript-go/shim/vfs/cachedvfs"
 	"github.com/microsoft/typescript-go/shim/vfs/osvfs"
 	rslintconfig "github.com/web-infra-dev/rslint/internal/config"
-	"github.com/web-infra-dev/rslint/internal/config/discovery"
 	"github.com/web-infra-dev/rslint/internal/config/target"
+	"github.com/web-infra-dev/rslint/internal/linter"
+	"github.com/web-infra-dev/rslint/internal/output"
+	"github.com/web-infra-dev/rslint/internal/program/loader"
+	"github.com/web-infra-dev/rslint/internal/rule"
+	"github.com/web-infra-dev/rslint/internal/rules"
 	"github.com/web-infra-dev/rslint/internal/term"
 )
-
-// lintArgs is the parsed CLI flag set, decoupled from the global flag
-// package so each entry point can build it: parseLintFlags from argv, and
-// runCLI additionally from the IPC init-handshake payload.
-type lintArgs struct {
-	Init           bool
-	Config         string
-	Fix            bool
-	TypeCheck      bool
-	TypeCheckOnly  bool
-	TraceOut       string
-	CpuprofOut     string
-	SingleThreaded bool
-	Format         string
-	NoColor        bool
-	ForceColor     bool
-	// StdoutIsTTY is the TTY fact for the real output destination, reported
-	// by the Node host via the IPC init payload (the Go process's own stdout
-	// is an IPC pipe and says nothing about the user's terminal). False when
-	// unavailable (for example in the wasm build).
-	StdoutIsTTY bool
-	Quiet       bool
-	Timing      bool
-	TimingLimit int
-	MaxWarnings int
-	StartTimeMs int64
-	RuleFlags   []string
-	// Positional args resolved into existing-dir vs file paths.
-	AllowFiles []string
-	AllowDirs  []string
-	// FS is an optional run-scoped filesystem shared with Go config discovery.
-	// Native CLI supplies one cached instance so the later target/program phases
-	// reuse directory entries already read by the staged frontier.
-	FS vfs.FS
-	// DeferTimingTable, when non-nil, receives the rendered --timing table
-	// instead of it being printed to stderr by the pipeline. The IPC entry
-	// uses it to emit the table only after the async stdout forwarding has
-	// drained, so the table cannot interleave with the lint report.
-	DeferTimingTable func(table string)
-	// ConfigCatalog is the immutable result of Go-owned JS/TS config discovery.
-	// A nil or empty automatic catalog selects the native JSON/JSONC loader;
-	// explicit catalogs remain authoritative even when their config is empty.
-	ConfigCatalog *discovery.ConfigCatalog
-}
-
-// repeatedFlag collects multiple values for the same flag (e.g. --rule used multiple times).
-type repeatedFlag []string
-
-func (f *repeatedFlag) String() string     { return strings.Join(*f, ", ") }
-func (f *repeatedFlag) Set(v string) error { *f = append(*f, v); return nil }
-
-const usage = `🚀 Rslint - Rocket Speed Linter
-
-Usage:
-  rslint [OPTIONS] [files...]
-
-Options:
-  --init                Initialize a default config in the current directory.
-  -c, --config PATH     Which rslint config file to use.
-  --format FORMAT       Output format: default | jsonline | github | gitlab
-  --fix                 Automatically fix problems
-  --type-check          Enable TypeScript type checking
-  --type-check-only     Run only TypeScript type checking (skip all lint rules)
-  --no-color            Disable colored output
-  --force-color         Force colored output
-  --quiet               Report errors only
-  --timing [all|N]      Print a per-rule timing table (all rules, or top N)
-  --max-warnings Int    Number of warnings to trigger nonzero exit code
-  --rule RULE           Rule override, e.g. 'no-console: error' (repeatable)
-  -h, --help            Show help
-`
-
-// groupDiagsByFile groups a flat slice of diagnostics by their source file name.
-func groupDiagsByFile(diags []rule.RuleDiagnostic) map[string][]rule.RuleDiagnostic {
-	m := make(map[string][]rule.RuleDiagnostic)
-	for _, d := range diags {
-		f := d.FilePath
-		m[f] = append(m[f], d)
-	}
-	return m
-}
-
-// remapDiagnosticTargetPaths keeps diagnostics in the caller's target path
-// space when a TypeScript Program represents that target by another lexical or
-// canonical source-file path. SourceFile remains unchanged because ranges and
-// fixes are defined against its text; FilePath controls display and disk writes.
-func remapDiagnosticTargetPaths(
-	diags []rule.RuleDiagnostic,
-	lintTargetBySourcePath map[string]target.File,
-	filesystems ...vfs.FS,
-) {
-	if len(lintTargetBySourcePath) == 0 {
-		return
-	}
-	var fsys vfs.FS
-	if len(filesystems) > 0 {
-		fsys = filesystems[0]
-	}
-	for i := range diags {
-		if target, ok := lookupLintTarget(lintTargetBySourcePath, diags[i].FilePath, fsys); ok {
-			diags[i].FilePath = target.Path
-		}
-	}
-}
-
-type typeScriptDiagnosticDedupeKey struct {
-	path     string
-	ruleName string
-	pos      int
-	end      int
-	message  string
-}
-
-// deduplicateTypeScriptDiagnostics joins the lint-target syntax path and the
-// program-wide type-check path. A file governed by a config without a tsconfig
-// can be parsed into a source-only rslint Program while also belonging to another
-// config's ts-go Program; --type-check legitimately visits both, but the same TypeScript
-// diagnostic must be reported once.
-func deduplicateTypeScriptDiagnostics(
-	diags []rule.RuleDiagnostic,
-	fsys vfs.FS,
-	preferredCallerTargets ...map[string]string,
-) []rule.RuleDiagnostic {
-	if len(diags) == 0 {
-		return diags
-	}
-	var callerTargetByCanonicalPath map[string]string
-	if len(preferredCallerTargets) > 0 {
-		callerTargetByCanonicalPath = preferredCallerTargets[0]
-	}
-	if len(diags) == 1 {
-		if diags[0].Origin == rule.DiagnosticOriginTypeScript {
-			canonicalID := canonicalFilesystemPathID(diags[0].FilePath, fsys)
-			if callerTarget := callerTargetByCanonicalPath[canonicalID]; callerTarget != "" {
-				diags[0].FilePath = callerTarget
-			}
-		}
-		return diags
-	}
-	bestIndex := make(map[typeScriptDiagnosticDedupeKey]int)
-	keys := make([]typeScriptDiagnosticDedupeKey, len(diags))
-	for i, diagnostic := range diags {
-		if diagnostic.Origin != rule.DiagnosticOriginTypeScript {
-			continue
-		}
-		key := typeScriptDiagnosticDedupeKey{
-			path:     canonicalFilesystemPathID(diagnostic.FilePath, fsys),
-			ruleName: diagnostic.RuleName,
-			pos:      diagnostic.Range.Pos(),
-			end:      diagnostic.Range.End(),
-			message:  diagnostic.Message.Description,
-		}
-		keys[i] = key
-		current, exists := bestIndex[key]
-		if !exists || preferTypeScriptDiagnostic(diagnostic, diags[current], callerTargetByCanonicalPath[key.path], fsys) {
-			bestIndex[key] = i
-		}
-	}
-
-	result := diags[:0]
-	for i, diagnostic := range diags {
-		if diagnostic.Origin != rule.DiagnosticOriginTypeScript {
-			result = append(result, diagnostic)
-			continue
-		}
-		key := keys[i]
-		if bestIndex[key] != i {
-			continue
-		}
-		if callerTarget := callerTargetByCanonicalPath[key.path]; callerTarget != "" {
-			diagnostic.FilePath = callerTarget
-		}
-		result = append(result, diagnostic)
-	}
-	return result
-}
-
-func preferTypeScriptDiagnostic(candidate rule.RuleDiagnostic, current rule.RuleDiagnostic, callerTarget string, fsys vfs.FS) bool {
-	if callerTarget != "" {
-		candidateIsCaller := exactFilesystemPathID(candidate.FilePath) == exactFilesystemPathID(callerTarget)
-		currentIsCaller := exactFilesystemPathID(current.FilePath) == exactFilesystemPathID(callerTarget)
-		if candidateIsCaller != currentIsCaller {
-			return candidateIsCaller
-		}
-	}
-	candidatePath := tspath.NormalizePath(candidate.FilePath)
-	currentPath := tspath.NormalizePath(current.FilePath)
-	if candidatePath != currentPath {
-		return candidatePath < currentPath
-	}
-	return false
-}
-
-// applyFixPass applies auto-fixes for all files in diagnosticsByFile,
-// writes fixed content to disk, and returns the number of issues fixed. Write
-// failures are returned after all independent files have been attempted.
-func applyFixPass(diagnosticsByFile map[string][]rule.RuleDiagnostic, fsys vfs.FS) (int, error) {
-	fixed := 0
-	fileNames := make([]string, 0, len(diagnosticsByFile))
-	for fileName := range diagnosticsByFile {
-		fileNames = append(fileNames, fileName)
-	}
-	sort.Strings(fileNames)
-
-	var writeErrors []error
-	for _, fileName := range fileNames {
-		fileDiagnostics := diagnosticsByFile[fileName]
-		var diagnosticsWithFixes []rule.RuleDiagnostic
-		for _, d := range fileDiagnostics {
-			if len(d.Fixes()) > 0 {
-				diagnosticsWithFixes = append(diagnosticsWithFixes, d)
-			}
-		}
-		if len(diagnosticsWithFixes) == 0 {
-			continue
-		}
-
-		// The parsed text has no byte order mark — decoding the file consumed
-		// it — so put it back before fixing. ApplyRuleFixes carries it through
-		// to the bytes written here, and drops it only for a fix that asks.
-		originalContent := diagnosticsWithFixes[0].SourceFile.Text()
-		if utils.SourceHasBOM(fsys, fileName) {
-			originalContent = utils.BOM + originalContent
-		}
-		fixedContent, unapplied, wasFixed := linter.ApplyRuleFixes(originalContent, diagnosticsWithFixes)
-
-		if wasFixed {
-			err := os.WriteFile(fileName, []byte(fixedContent), 0644)
-			if err != nil {
-				writeErrors = append(writeErrors, fmt.Errorf("write fixed file %q: %w", fileName, err))
-			} else {
-				fixed += len(diagnosticsWithFixes) - len(unapplied)
-			}
-		}
-	}
-	return fixed, errors.Join(writeErrors...)
-}
-
-// allowFileWarningKind categorizes why a CLI-specified file won't have lint
-// rules applied. These are Phase-1 (lint) concepts; Phase 2 (type-check) is
-// not affected by either case (it runs program-wide regardless of CLI scope
-// and rslint ignores).
-type allowFileWarningKind int
-
-const (
-	allowFileNotFound allowFileWarningKind = iota
-	allowFileIgnored
-)
-
-// allowFileWarning is the structured form of a "this CLI-specified file
-// won't be linted" diagnostic. Returned by collectAllowFileWarnings so the
-// emission policy (skip in --type-check-only, format with relative paths,
-// etc.) can be unit-tested without capturing stderr.
-type allowFileWarning struct {
-	Path string // absolute, normalized — matches what was put in allowFiles
-	Kind allowFileWarningKind
-}
-
-// formatAllowFileWarning renders an allowFileWarning for stderr emission,
-// resolving the absolute path against comparePathOptions. Returns "" for
-// unknown kinds so callers can safely ignore the empty case.
-func formatAllowFileWarning(w allowFileWarning, opts tspath.ComparePathsOptions) string {
-	rel := tspath.ConvertToRelativePath(w.Path, opts)
-	switch w.Kind {
-	case allowFileNotFound:
-		return fmt.Sprintf("warning: %s was not found, skipping", rel)
-	case allowFileIgnored:
-		return fmt.Sprintf("warning: %s is ignored because of a matching ignore pattern", rel)
-	}
-	return ""
-}
-
-// collectAllowFileWarnings maps target-plan explicit-file outcomes to CLI
-// messages. It performs no filesystem or config lookup: target selection has
-// already frozen the file identity, owner, ignore decision, and existence.
-// Program membership is deliberately not consulted because a file outside
-// every tsconfig can still be linted.
-//
-// This is a Phase-1 concern only. In --type-check-only mode the lint phase
-// is skipped, so callers MUST gate emission on `!typeCheckOnly` — otherwise
-// users see misleading messages like "is ignored because of a matching
-// ignore pattern" while Phase 2 happily reports type errors for that same
-// file. See website/docs/en/guide/type-checking.md ("type-check is not
-// constrained by `files`/`ignores`").
-func collectAllowFileWarnings(
-	outcomes []target.ExplicitFileOutcome,
-) []allowFileWarning {
-	if len(outcomes) == 0 {
-		return nil
-	}
-
-	var out []allowFileWarning
-	for _, outcome := range outcomes {
-		if outcome.Ignored {
-			out = append(out, allowFileWarning{Path: outcome.Path, Kind: allowFileIgnored})
-			continue
-		}
-		if !outcome.Exists {
-			out = append(out, allowFileWarning{Path: outcome.Path, Kind: allowFileNotFound})
-			continue
-		}
-	}
-	return out
-}
-
-// shouldShortCircuitOutput returns true when rslint should bail early
-// without printing diagnostics or a summary. The short-circuit exists so
-// that e.g. `rslint nonexistent-file.ts` returns 0 with no spurious output
-// when Phase 1 visited zero files.
-//
-// Any type-check mode (`--type-check` or `--type-check-only`) must NOT take
-// the short-circuit: Phase 2 runs program-wide and is not gated by the CLI
-// Scope/PerProgramFilter that drives lintedFileCount, so lintedFileCount==0
-// is a normal state in which Phase 2 may still have produced diagnostics.
-// Short-circuiting there would silently drop type errors that the user
-// explicitly asked for — see website/docs/en/guide/type-checking.md.
-func shouldShortCircuitOutput(typeCheckOnly, typeCheck, scopeRestricted bool, lintedFileCount int32) bool {
-	if typeCheckOnly || typeCheck {
-		return false
-	}
-	return scopeRestricted && lintedFileCount == 0
-}
-
-// validateTypeCheckOnlyFlags rejects --type-check-only combined with flags
-// whose semantics depend on the lint phase that this mode disables. Returns
-// (0, "") when the combination is valid (or --type-check-only isn't set);
-// otherwise returns (exitCode > 0, stderr message). Pulled out as a pure
-// function so the policy can be exercised in unit tests.
-func validateTypeCheckOnlyFlags(typeCheckOnly, fix bool, ruleFlags []string) (int, string) {
-	if !typeCheckOnly {
-		return 0, ""
-	}
-	if fix {
-		return 2, "error: --fix cannot be combined with --type-check-only (no lint rules run, nothing to fix)"
-	}
-	if len(ruleFlags) > 0 {
-		return 2, "error: --rule cannot be combined with --type-check-only (no lint rules run)"
-	}
-	return 0, ""
-}
 
 func isBroadProjectLoadScope(
 	allowFiles []string,
@@ -397,82 +48,6 @@ func isBroadProjectLoadScope(
 		}
 	}
 	return false
-}
-
-func cloneConfigMap(configMap map[string]rslintconfig.RslintConfig) map[string]rslintconfig.RslintConfig {
-	if configMap == nil {
-		return nil
-	}
-	cloned := make(map[string]rslintconfig.RslintConfig, len(configMap))
-	for dir, cfg := range configMap {
-		cloned[dir] = slices.Clone(cfg)
-	}
-	return cloned
-}
-
-func configsForOwners(
-	configMap map[string]rslintconfig.RslintConfig,
-	owners []string,
-) map[string]rslintconfig.RslintConfig {
-	if len(configMap) == 0 || len(owners) == 0 {
-		return nil
-	}
-	active := make(map[string]rslintconfig.RslintConfig, len(owners))
-	for _, owner := range owners {
-		if entries, ok := configMap[owner]; ok {
-			active[owner] = entries
-		}
-	}
-	return active
-}
-
-// validateResolvedRuleOptions runs rule-options schema validation over the
-// resolved configuration: every configMap config in multi-config mode (each
-// message suffixed with the owning config directory, since the same rule can
-// be misconfigured differently per config), or the single rslintConfig
-// otherwise. It returns normalized copies plus one formatted message per
-// failure, leaving the input configs untouched.
-func validateResolvedRuleOptions(
-	configMap map[string]rslintconfig.RslintConfig,
-	rslintConfig rslintconfig.RslintConfig,
-	catalog *rule.Catalog,
-) (map[string]rslintconfig.RslintConfig, rslintconfig.RslintConfig, []string) {
-	var messages []string
-	if configMap == nil {
-		normalized, optionsErrors := rslintconfig.ValidateRuleOptions(rslintConfig, catalog)
-		for _, optionsError := range optionsErrors {
-			messages = append(messages, optionsError.Error())
-		}
-		return nil, normalized, messages
-	}
-	configDirs := make([]string, 0, len(configMap))
-	for dir := range configMap {
-		configDirs = append(configDirs, dir)
-	}
-	slices.Sort(configDirs)
-	normalizedMap := make(map[string]rslintconfig.RslintConfig, len(configMap))
-	for _, dir := range configDirs {
-		normalized, optionsErrors := rslintconfig.ValidateRuleOptions(configMap[dir], catalog)
-		normalizedMap[dir] = normalized
-		for _, optionsError := range optionsErrors {
-			messages = append(messages, fmt.Sprintf("%s (config at %s)", optionsError.Error(), dir))
-		}
-	}
-	return normalizedMap, rslintConfig, messages
-}
-
-func deriveRuleCatalog(plugins []rslintconfig.EslintPluginEntry) (*rule.Catalog, []string) {
-	return rules.All().ForESLintPlugins(plugins)
-}
-
-func reportShadowedPluginRules(shadowed []string) {
-	for _, ruleName := range shadowed {
-		fmt.Fprintf(
-			os.Stderr,
-			"rslint: plugin rule %q is shadowed by a built-in rule of the same name; using the built-in.\n",
-			ruleName,
-		)
-	}
 }
 
 // resolveStartTime returns the start time for timing output.
@@ -522,104 +97,6 @@ func loadGitignoreAndProjects(
 		return config, loader.ProjectSet{}, programErr
 	}
 	return configWithIgnores, projects, nil
-}
-
-// parseLintFlags parses the lint CLI flags out of argv into a lintArgs.
-// It uses a fresh FlagSet (not the global flag.CommandLine) so it is
-// callable more than once per process, and ContinueOnError so a bad flag
-// returns a fatal exit code instead of os.Exit-ing past caller cleanup.
-// A non-zero fatalExitCode means the caller should return it immediately
-// (the diagnostic was already printed to stderr).
-func parseLintFlags(argv []string) (args lintArgs, help bool, fatalExitCode int) {
-	fs := flag.NewFlagSet("rslint", flag.ContinueOnError)
-	fs.SetOutput(os.Stderr)
-	fs.Usage = func() { fmt.Fprint(os.Stderr, usage) }
-
-	var ruleFlags repeatedFlag
-
-	fs.StringVar(&args.Format, "format", "default", "output format")
-	fs.StringVar(&args.Config, "config", "", "which rslint config to use")
-	fs.StringVar(&args.Config, "c", "", "which rslint config to use")
-	fs.BoolVar(&args.Init, "init", false, "initialize a default config in the current directory")
-	fs.BoolVar(&args.Fix, "fix", false, "automatically fix problems")
-	fs.BoolVar(&args.TypeCheck, "type-check", false, "enable TypeScript type checking")
-	fs.BoolVar(&args.TypeCheckOnly, "type-check-only", false, "run only TypeScript type checking (skip all lint rules)")
-	fs.BoolVar(&help, "help", false, "show help")
-	fs.BoolVar(&help, "h", false, "show help")
-	fs.BoolVar(&args.NoColor, "no-color", false, "disable colored output")
-	fs.BoolVar(&args.ForceColor, "force-color", false, "force colored output")
-	fs.BoolVar(&args.Quiet, "quiet", false, "report errors only")
-	var timingValue string
-	fs.StringVar(&timingValue, "timing", "", "print a per-rule timing table: 'all' or a top rule count")
-	fs.IntVar(&args.MaxWarnings, "max-warnings", -1, "Number of warnings to trigger nonzero exit code")
-
-	fs.StringVar(&args.TraceOut, "trace", "", "file to put trace to")
-	fs.StringVar(&args.CpuprofOut, "cpuprof", "", "file to put cpu profiling to")
-	fs.BoolVar(&args.SingleThreaded, "singleThreaded", false, "run in single threaded mode")
-	fs.Int64Var(&args.StartTimeMs, "start-time", 0, "internal: epoch milliseconds from Node.js entry point")
-	fs.Var(&ruleFlags, "rule", "rule override, e.g. 'no-console: error' (repeatable)")
-
-	if err := fs.Parse(argv); err != nil {
-		// ContinueOnError: fs already printed the diagnostic to stderr.
-		return args, help, 2
-	}
-	args.RuleFlags = []string(ruleFlags)
-
-	// The Node.js entry point fills in the default "all" when the user
-	// passes a bare --timing, so the value is always present here.
-	switch {
-	case timingValue == "":
-	case strings.EqualFold(timingValue, "all"):
-		args.Timing = true
-	default:
-		n, err := strconv.Atoi(timingValue)
-		if err != nil || n <= 0 {
-			fmt.Fprintf(os.Stderr, "invalid value %q for flag -timing: expected \"all\" or a positive rule count\n", timingValue)
-			return args, help, 2
-		}
-		args.Timing = true
-		args.TimingLimit = n
-	}
-
-	// --type-check-only implies --type-check and skips all lint rules.
-	// Reject incompatible flag combinations before doing any work.
-	if code, msg := validateTypeCheckOnlyFlags(args.TypeCheckOnly, args.Fix, args.RuleFlags); code != 0 {
-		fmt.Fprintln(os.Stderr, msg)
-		return args, help, code
-	}
-	if args.TypeCheckOnly {
-		args.TypeCheck = true
-	}
-
-	// Collect file/directory arguments for targeted linting (e.g. rslint file1.ts src/)
-	if positional := fs.Args(); len(positional) > 0 {
-		for _, arg := range positional {
-			absPath, err := filepath.Abs(arg)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "error resolving path %s: %v\n", arg, err)
-				return lintArgs{}, help, 1
-			}
-			// NOTE: we intentionally do NOT call filepath.EvalSymlinks here.
-			// EvalSymlinks resolves symlinks (macOS /tmp → /private/tmp) and
-			// Windows 8.3 short names to long names, but the rest
-			// of the pipeline (os.Getwd, TypeScript file names, configDir)
-			// uses unresolved CWD-based paths. Resolving only file args would
-			// create a format mismatch causing failures in lint-target detection,
-			// config matching, dir scoping, and gitignore checks.
-			// Edge cases (e.g. user passes a symlink-resolved absolute path)
-			// are handled by isFileAllowed's os.SameFile fallback in linter.go
-			// and program file index realpath aliases during target binding.
-			normalized := tspath.NormalizePath(absPath)
-			info, statErr := os.Stat(absPath)
-			if statErr == nil && info.IsDir() {
-				args.AllowDirs = append(args.AllowDirs, normalized)
-			} else {
-				args.AllowFiles = append(args.AllowFiles, normalized)
-			}
-		}
-	}
-
-	return args, help, 0
 }
 
 // executeLintPipeline runs the full lint flow (config load → program build →
